ARROW-4017: [C++] Move vendored libraries in dedicated directory
Also update mapbox::variant to v1.1.5 (I'm not sure which version was previously vendored).
