+++ cpp/build-support/lint_cpp_cli.py
@@ -70,13 +70,10 @@ def lint_file(path):
 
 EXCLUSIONS = [
     'arrow/python/iterators.h',
-    'arrow/util/date.h',
     'arrow/util/hashing.h',
     'arrow/util/macros.h',
     'arrow/util/parallel.h',
-    'arrow/util/string_view/string_view.hpp',
-    'arrow/util/xxhash/xxhash.c',
-    'arrow/util/xxhash/xxhash.h',
+    'arrow/vendored',
     'arrow/visitor_inline.h',
     'gandiva/cache.h',
     'gandiva/jni',
